arguments for read-write native queries (#16)
Hooks up native query arguments, but only for mutations for this PR. Converts JSON argument values to BSON according to the native query's parameter types. Implements a system to interpolate argument values into the BSON command document sent to MongoDB.

I'm going to handle arguments for read-only native queries in a follow-up PR since I realized that there is still more to do to resolve arguments in the presence of query variables.

It turns out that the earlier change I made to switch configuration fields to maps broke YAML deserialization due to a bug described here: https://github.com/dtolnay/serde-yaml/issues/344. So I've converted the native query fixtures to JSON for the time being. We'll have to either find a workaround for that bug, or disable YAML parsing. I'm inclined to do a bit of work to try for a workaround since I think the YAML format is nicer for working with: it's much less verbose, and I had to delete a comment I had in one fixture because JSON.

Ticket: https://hasurahq.atlassian.net/browse/MDB-87