rename "native query" to "native procedure"; remove read-only native procedures (#29)
Going forward "native query" will refer to a thing defined by an aggregation pipeline (not an arbitrary MongoDB command) that will appear in the data graph as a function. (Maybe we will add support for native queries that appear as collections at some point, but we've been asked to specifically implement the function case.)

What was previously called a "native query" is now called a "native procedure". It maps exclusively to a procedure in the data graph. The read-write/read-only switch is gone - procedures will only be used for mutations.

https://hasurahq.atlassian.net/browse/MDB-102