prevent response serialization from assuming objects are relations (#34)
The change in #27 fixed serializing query responses with object values, but serializing foreach responses stopped working. This PR switches back to roughly the logic in which foreach responses were working, and fixes the issue with object values in a different way.

This is a quick patch - we still need to rework response serialization to use `bson_to_json` translation. That will require using the query request to direct response serialization which should make everything more robust.

I removed the use of `JsonResponse` in this PR because it serves no purpose for this connector. `JsonResponse` is useful for connectors that get a JSON response from the database, which the connector can pass through without processing. But as far as we can determine MongoDB always sends BSON so to get JSON output we're stuck doing response processing in the connector.

There's also a fix in here to avoid using `$literal` expression escaping in a context where expressions are not evaluated.

Fixes https://hasurahq.atlassian.net/browse/MDB-19

---------

Co-authored-by: Brandon Martin <brandon@codedmart.com>