switch default arion project to use sample_mflix db instead of chinook (#35)
This means that when you run `arion up -d` you get the sample_mflix database. This database is better for document-oriented testing because it includes object and array values within documents.

Relationships and native procedures are not configured yet.

I started setting things up so we can use both databases via two connector instances. We'll have to set up a supergraph with namespaces if we want to finish that.

The ndc-test and e2e test arion projects still use chinook.