fix order of results for query requests with more than 10 variable sets (#37)
## Describe your changes

In query requests with more than 10 variable sets results were sent back in the wrong order because facet fields were put into a `BTreeMap` which sorted them lexicographically. This PR switches to a `Vec` of tuples to preserve expected order.

While working on this I updated the dev services to run two connectors with a remote relationship configured. If you run `arion up -d` you can access GraphiQL at http://localhost:7100/ and test the fix with a query like this one:

```graphql
query AlbumMovies {
  album(limit: 11) {
    title
    movies(limit: 2) {
      title
      runtime
    }
    albumId
  }
}
```

The remote relationship matches `albumId` in the `Album` collection of the chinook database with the `runtime` field in the `movies` collection of the sample_mflix database. It's not a sensical relationship but it gives us something to test with since both of those fields are ints with plenty of values that line up.

## Issue ticket number and link

[MDB-111](https://hasurahq.atlassian.net/browse/MDB-111)