move v3-e2e-testing fetcher out of flake (#41)
Removes the `v3-e2e-testing` input from `flake.nix`, and moves it to a fetcher in the arion configuration that runs the e2e service. The e2e repo is private - if we keep it in the flake the entire nix setup will be broken for anyone who isn't set up to access the e2e repo with ssh authentication. Moving the fetcher to that specific arion config file prevents an authentication error unless you're specifically running e2e tests.

The advantage of having inputs in `flake.nix` is that we can update versions with commands like `nix flake update`, and nix handles locking versions automatically. OTOH with `builtins.fetchGit` we have to edit the file to set the git revision that we want.