serialize query responses according to schema types (#53)
This changes response serialization to use simple JSON instead of Extended JSON (except in cases of fields with the type `ExtendedJSON`).

Mutation responses have not been updated with this change yet - that will be in a follow-up PR.

This could be greatly simplified if we had an internal query request type that used copies of object type definitions in all places instead of object type names. I've got some thoughts on changing `v3_to_v2_query_request` to work with an internal type like that instead of with the v2 query request type. But that's a big change that will have to wait until later.