test query request with one variable set (#69)
Add a test just to make sure this case works. I thought I remembered MongoDB rejecting `$facet` stages with only one sub-pipeline, but according to this test that case works just fine in MongoDB versions 5, 6, and 7.

This PR expands the integration tests to allow sending a query request directly to a connector, bypassing the engine. It also cleans up some of the test helper code. The test snapshots all needed to be updated because I changed a function name, which changed the expression used to produce snapshot results, and the expression is captured as part of the snapshot file.