refactor: remove v2 types (#71)
Removes v2 types, and replaces them with a new set of internal types that closely match ndc models. The root of the new types is called `QueryPlan` instead of `QueryRequest`. It is a denormalized version of a query request. In particular `QuerPlan`

- puts type annotations on fields, variable references, etc.
- uses a `Type` type that inlines object type references
- collects relationship reference details from all fields, filters, and sorts and places relevant details in one  map in each sub-query (instead of having one relations map at the `QueryRequest` level, the new types have a map on each `Query` value for relations with that query's collection as a source, and that are referenced in that query)
- collects one map of join details at the `QueryPlan` level (the analog of `QueryRequest`) for existence checks on unrelated collections

The changes simplify emitting MongoDB aggregation code, and lower the impedence mismatch to using features of v3 that aren't reflected in v2.

I preserved and updated all the existing tests except one that tests filtering by fields of a nested object of a related collection which is something that worked in v2, but will require an update to the latest ndc-spec to work in v3.