support root collection column references (#75)
Supports root collection column references correctly. We were using `$$ROOT` which just refers to the current collection. Root references actually reference the collection of the nearest query up the tree in the query request. The query plan that we produce from a query request is structured differently, and has `Query` values in positions where the request doesn't. So this PR introduces a "scope" concept in the query plan to capture the intended reference scopes from the request.

A couple of notes here on how the MongoDB expressions are evaluated:

The special variable `$$ROOT` refers to the "current document" being processed by the pipeline. When we look up a relation we use a `$lookup` stage which has it's own pipeline. If we reference `$$ROOT` _inside_ if the `$lookup` pipeline then it resolves to documents in the related collection - the one we are looking up.

On the other hand the `$lookup` stage also accepts a `let` map to bind variables. When we reference `$$ROOT` in a variable assignment in this `let` map it references the collection we are joining **from**, not the collection we are joining to. And then the variable bound in the `let` map is in scope in the `$lookup` pipeline.

[MDB-6](https://hasurahq.atlassian.net/browse/MDB-6)