fix: count aggregates return 0 instead of null if no rows match (#85)
Fix for count aggregates in cases where the query does not match any rows. Previously the connector returned null which is an appropriate response for other aggregations (like sum), but not for counts. This change fixes the problem by substituting zero for null for count aggregates only.

This is a port of a fix for the same issue from the v2 agent.

Ticket: [GDC-1345](https://hasurahq.atlassian.net/browse/GDC-1345)