ci: update rust-sec advisory db in every test run (#100)
We have had a security vulnerability check on cargo dependencies in place for a long time. But it's necessary to update the advisory db to get the latest advisories. This change updates a github workflow to run the update on every CI test run.

I tested that the check catches problems by running a test with `gix-fs` installed at version 0.10.2. See https://rustsec.org/advisories/RUSTSEC-2024-0350.html