ci: run integration tests with MongoDB v8; remove MongoDB v5 from testing (#112)
MongoDB v8 has a new stable release so it's past time to include it in our tests.

MongoDB v5 is EOL this month, and it doesn't support our remote join implementation anyway so it's been off our supported versions list for a while. We've been having to skip a bunch of tests in that version, and now we don't have to do that anymore.