escape field names in relation mappings when necessary (#113)
Previously on attempts to join on field names that require escaping the request would fail with an error saying that a name was not valid. This change does the necessary escaping to make the join work.

This is the last change that was needed to get rid of `safe_name` which was the quick-and-easy fail-on-problematic-names solution that we used early on.