support $project stage in native query pipeline type inference (#126)
This is work an an in-progress feature that is gated behind a feature flag, `native-query-subcommand`.

This change allows the native query configuration generator to infer the output type of [`$project`](https://www.mongodb.com/docs/manual/reference/operator/aggregation/project/) stages in pipelines. It also enables inference for parameters used in `$limit` and `$skip` stages.

`$project` is a complicated feature, but it's important to support it because it's widely used. It allows adding, removing, or modifying document fields in the document-processing pipeline. It has two modes: it can either remove fields, or it can select a subset of fields to keep while optionally adding more or changing values of kept fields. It also has dotted-path notation for easily manipulating nested structures.