parse parameter type annotations when generating native query configuration (#128)
This is work an an in-progress feature that is gated behind a feature flag, `native-query-subcommand`.

Parses type annotations in placeholders that reference native query parameters. We already have an error message suggesting doing this, now the system actually reads these. For example,

```json
{
  "$match": {
    "imdb.rating": { "$gt": "{{ min_rating | int! }}" }
  }
}
```

The generated query for that configuration includes a parameter named `min_rating` with type `int`. Without the annotation it would have inferred `double`.

Type annotations are checked against the inferred type for the position so you will see errors if the annotated type is not compatible. Parameters are treated as contravariant so you can only provide a subtype of the inferred type - for example you can constrain a parameter type to be non-nullable even if it is in a nullable context. There are cases where the type checker cannot infer a type in which case annotations are necessary.

I know we already have a similar parser for type expressions in hml files. I thought it would be easier to write a new one specialized for this connector and for MongoDB scalar types since it's about 100 LOC.

Type expressions use GraphQL syntax to match types as seen in GraphQL and in hml. There is one addition: I invented a syntax for predicate types, `predicate<ObjectTypeName>`. The parser happens to be written so that if the angle brackets are absent the word `predicate` will be interpreted as an object type so we don't have a problem if a user want to use an object type named "predicate".

On the other hand this parser does not allow object type names that match MongoDB scalar type names.

While I was working on this I noticed some issues with unifying types in the presence of nullability, and with displaying errors when a type annotation can't be unified with inferred constraints for a parameter's context. So I included some fixes in those areas.