recreate supergraph fixture using latest cli version (#134)
Recreates the supergraph in `fixtures/hasura/` using ddn cli v2.15.0 with as little hand-editing of metadata as possible. This makes it much easier to continue updating metadata using the cli.

This PR mainly affects the integration test and local development testing setup. But since I needed to update the GraphQL Engine version used in integration tests I updated flake inputs. That leads to updating the openssl version. And since all of that leads to rebuilding stuff anyway I also updated the Rust toolchain to v1.83.0.

An issue that came up is that the hack I had tested to filter by array fields does not seem to work anymore. It looks like this is because of a change in the engine - the relevant test fails with the previous metadata fixture with the latest engine version. I've disabled the test. This is the test query:

```graphql
query {
  movies(where: { cast: { _eq: "Albert Austin" } }) {
    title
    cast
  }
}
```

Note that `cast` is an array of strings. It looks like we may need to wait for ndc-spec v0.2 to be able to do this properly.

The changes here include a lot of automatically-generated configuration in `fixtures/hasura/` so it's a big set of changes.