implement aggregations on nested fields (#136)
Turns on the nested field aggregation connector capability, and makes it work.

This also touches up the aggregation implementation generally. It uses the updated, unified system for referencing fields with paths. It also changes aggregation result types (besides count) to be nullable which avoids an error when aggregating over an empty document set.