fix deploy automation (#142)
Since the last release the `ubuntu-latest` runner for this repo switched from `ubuntu-22.04` to `ubuntu-24.04`. That came with changes to docker permissions configuration that broke the `deploy::docker` job. The problem is the same one described here:

https://github.com/actions/runner-images/issues/10443

Except that instead of `skopeo copy` we hit the error running `buildah`.

I've worked around the problem by rolling the `deploy::docker` job back to the `ubuntu-22.04` runner. To head of similar future issues I've changed the runners for all other jobs to a fixed runner version.