add uuid scalar type (#148)
UUIDs are stored in BSON using the `binData` type which we already support. But the `binData` type is not especially user friendly to work with in its generalized form, and UUIDs are a common use case. It is helpful to surface UUIDs as a first-class scalar type that serializes to a string.

The `binData` type has a numeric `subType` field to give a hint as to how the stored binary data should be interpreted. There are two subtypes for UUIDs, 3 & 4, but subtype 4 is indicated as the "current" representation for UUIDs. The `UUID()` constructor in `mongosh` produces `binData` values with subtype 4.

This change:
- adds a scalar type called `UUID`
- sets the NDC type representation for `UUID` to `String`
- converts `UUID` values to BSON as `binData` values with subtype 4 - inputs are parsed from string representations using functions provided by the BSON library that is bundled with the MongoDB Rust driver
- serializes `UUID` values to JSON as strings using provided BSON library functions - for example, `4ca4b7e7-6f6a-445b-b142-9d6d252d92bc`
- updates introspection to infer the type `UUID` when encountering fields with `binData` values with subtype 4
- introspection infers the type `BinData` instead for fields where subtype 4 occurs alongside `binData` values with other subtypes

This allows for more user-friendly queries involving UUIDs. For example filtering by UUID changed from this:

```gql
id: { _eq: { base64: "TKS3529qRFuxQp1tJS2SvA==" subType: "04" } }
```

and now looks like this:

```gql
id: { _eq: "4ca4b7e7-6f6a-445b-b142-9d6d252d92bc" }
```