implement group by for single-column aggregates (#144)
Implements most of the functionality for the capability `query.aggregates.group_by`. There are still a couple of things to follow up on.

Counts are not implemented for group by queries yet. I'll follow up on those in [ENG-1568](https://linear.app/hasura/issue/ENG-1568/[mongodb]-implement-count-for-group-by). (Counts are implemented in #145 which can be merged after this PR is merged.)

There is a bug involving multiple references to the same relationship that should be resolved. I'll follow up in [ENG-1569](https://linear.app/hasura/issue/ENG-1569).

While working on this I removed the custom "count" aggregation - it is redundant, and I've been meaning to do that for a while. Users can use the standard count aggregations instead.

There is a change in here that explicitly converts aggregate result values for "average" and "sum" aggregations to the result types declared in the schema. This is necessary to avoid errors in response serialization for groups when aggregating over 128-bit decimal values. I applied the same type conversion for group and for root aggregates for consistency. This does mean there will be some loss of precision in those cases. But it also means we won't get back a JSON string in some cases, and a JSON number in others.