when filtering related collection need to include nested fields in comparison (#171)
This comes up when filtering on a nested field in a related collection. For example, where there is a relation from `comments` to `movie`:

```graphql
query {
  comments(where: {movie: {imdb: {rating: {_eq: 9.2}}}}) {
    text
    movie {
      title
      year
    }
  }
}
```

In this example no records are returned because `imdb.rating` is not in scope when during filtering where that `where` predicate is applied.

We already have logic in place to include columns in joins if they are needed for an expression. But that logic did not account for **nested** fields. This PR updates the query planner to recursively include nested fields as needed if they are used in expressions.