escape names with invalid characters in mongodb query field selection (#175)
We had a bug when selecting fields with names containing dollar signs or
dots, or when selecting fields using aliases containing those
characters. What happened is the mongodb query plan would produce a
stage like this:

```json
{
  "$replaceWith": {
    "foo.bar": "$title"
  }
}
```

The use of `"foo.bar"` leads to this error:

> MongoServerError: FieldPath field names may not contain '.'. Consider
using $getField or $setField.

The fix is, as the error suggests, to use the `$setField` operator which
takes a `field` argument that is not evaluated as an expression, and is
therefore allowed to contain arbitrary characters.

```json
{
  "$replaceWith": {
    "$setField": {
      "field": "foo.bar",
      "value": "$title",
      "input": "$$CURRENT"
    }
  }
}
```

But that leads to an issue when we want to select multiple fields: in
the first snippet the argument to `$replaceWith` was an object literal
where we could set multiple fields. OTOH `{ "$setField": { ... } }` is
an expression that produces an object with a single field, and does not
accept more keys. To select multiple fields with escaping we have to
nest calls to `$setField` for every selected field. For example,

```json
{
  "$replaceWith": {
    "$setField": {
      "field": "foo.bar",
      "value": "$title",
      "input": {
        "$setField": {
          "field": "year",
          "value": "$year",
          "input": {
            "$setField": {
              "field": "rated",
              "value": "$rated",
              "input": "$$CURRENT"
            }
          }
        }
      }
    }
  }
}
```

So that is what this PR does.